requirements.txt
python-dotenv
sentence-transformers
pinecone
groq
streamlit